+++ QLNS_SGU/QLNS_SGU/Presenter/LoaiNganhPresenter.cs
@@ -37,7 +37,7 @@ namespace QLNS_SGU.Presenter
     {
         private LoaiNganhForm _view;
         private bool checkNewRowExist = true;
-
+        UnitOfWorks unitOfWorks = new UnitOfWorks(new QLNSSGU_1Entities());
         public object UI => _view;
         public LoaiNganhPresenter(LoaiNganhForm view) => _view = view;
 
@@ -51,7 +51,6 @@ namespace QLNS_SGU.Presenter
         private void LoadDataToGrid()
         {
             SplashScreenManager.ShowForm(_view, typeof(WaitForm1), true, true, false, 0);
-            UnitOfWorks unitOfWorks = new UnitOfWorks(new QLNSSGU_1Entities());
             BindingList<LoaiNganh> listLoaiNganh = new BindingList<LoaiNganh>(unitOfWorks.LoaiNganhRepository.GetListLoaiNganh());
             _view.GCLoaiNganh.DataSource = listLoaiNganh;
             SplashScreenManager.CloseForm(false);
@@ -77,15 +76,33 @@ namespace QLNS_SGU.Presenter
         public void MouseDoubleClick(object sender, MouseEventArgs e)
         {
             GridHitInfo hinfo = _view.GVLoaiNganh.CalcHitInfo(e.Location);
-            if (hinfo.Column == _view.GVLoaiNganh.Columns[1])
+            int idRowFocused = Convert.ToInt32(_view.GVLoaiNganh.GetFocusedRowCellDisplayText("idLoaiNganh"));
+            if (!unitOfWorks.LoaiNganhRepository.CheckExistById(idRowFocused))
             {
-                _view.GVLoaiNganh.Columns[1].OptionsColumn.AllowEdit = true;
-                _view.GVLoaiNganh.ShowEditor();
+                if (hinfo.Column == _view.GVLoaiNganh.Columns[0])
+                {
+                    _view.GVLoaiNganh.Columns[0].OptionsColumn.AllowEdit = true;
+                    _view.GVLoaiNganh.ShowEditor();
+                }
+                if (hinfo.Column == _view.GVLoaiNganh.Columns[1])
+                {
+                    _view.GVLoaiNganh.Columns[1].OptionsColumn.AllowEdit = true;
+                    _view.GVLoaiNganh.ShowEditor();
+                }
             }
+            if(unitOfWorks.LoaiNganhRepository.CheckExistById(idRowFocused))
+            {
+                if (hinfo.Column == _view.GVLoaiNganh.Columns[1])
+                {
+                    _view.GVLoaiNganh.Columns[1].OptionsColumn.AllowEdit = true;
+                    _view.GVLoaiNganh.ShowEditor();
+                }
+            }            
         }
 
         public void HiddenEditor(object sender, EventArgs e)
         {
+            _view.GVLoaiNganh.Columns[0].OptionsColumn.AllowEdit = false;
             _view.GVLoaiNganh.Columns[1].OptionsColumn.AllowEdit = false;
         }
 
@@ -102,19 +119,31 @@ namespace QLNS_SGU.Presenter
         public void SaveData()
         {
             _view.GVLoaiNganh.CloseEditor();
-            _view.GVLoaiNganh.UpdateCurrentRow();
-            UnitOfWorks unitOfWorks = new UnitOfWorks(new QLNSSGU_1Entities());
+            _view.GVLoaiNganh.UpdateCurrentRow();            
             int row_handle = _view.GVLoaiNganh.FocusedRowHandle;
             int idRowFocused = Convert.ToInt32(_view.GVLoaiNganh.GetFocusedRowCellDisplayText("idLoaiNganh"));
-            if (idRowFocused == 0)
+            if (!unitOfWorks.LoaiNganhRepository.CheckExistById(idRowFocused))
             {
+                string idloainganh = _view.GVLoaiNganh.GetFocusedRowCellDisplayText("idLoaiNganh").ToString();
+                int id = -1;
+                if (!string.IsNullOrEmpty(idloainganh))
+                {
+                    id = Convert.ToInt32(idloainganh);
+                }
+                else
+                    XtraMessageBox.Show("Bạn chưa nhập Mã loại ngành.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 string loainganh = _view.GVLoaiNganh.GetFocusedRowCellDisplayText("tenLoaiNganh").ToString();
                 if (loainganh != string.Empty)
-                {                    
-                    unitOfWorks.LoaiNganhRepository.Create(loainganh);
-                    LoadDataToGrid();
-                    XtraMessageBox.Show("Lưu dữ liệu thành công.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    _view.GVLoaiNganh.MoveLast();
+                {       
+                    if (!unitOfWorks.LoaiNganhRepository.CheckExistById(id))
+                    {
+                        unitOfWorks.LoaiNganhRepository.Create(id, loainganh);
+                        LoadDataToGrid();
+                        XtraMessageBox.Show("Lưu dữ liệu thành công.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        _view.GVLoaiNganh.MoveLast();
+                    }
+                    else
+                        XtraMessageBox.Show("Mã loại ngành đã tồn tại. Vui lòng nhập mã khác.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
                 else
                 {
@@ -129,7 +158,7 @@ namespace QLNS_SGU.Presenter
                 if (loainganh != string.Empty)
                 {
                     unitOfWorks.LoaiNganhRepository.Update(id, loainganh);
-                    XtraMessageBox.Show("Lưu dữ liệu thành công.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);                    
+                    XtraMessageBox.Show("Lưu dữ liệu thành công.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
                 else
                 {
@@ -143,32 +172,28 @@ namespace QLNS_SGU.Presenter
         public void InitNewRow(object sender, InitNewRowEventArgs e)
         {            
             GridView gridView = sender as GridView;
+            gridView.SetRowCellValue(e.RowHandle, gridView.Columns[0], -1);
             gridView.SetRowCellValue(e.RowHandle, gridView.Columns[1], string.Empty); 
         }
 
         public void DeleteRow()
         {
-            try
+            int row_handle = _view.GVLoaiNganh.FocusedRowHandle;
+            if (row_handle >= 0)
             {
-                UnitOfWorks unitOfWorks = new UnitOfWorks(new QLNSSGU_1Entities());
-                int row_handle = _view.GVLoaiNganh.FocusedRowHandle;
-                if(row_handle >= 0)
+                int id = Convert.ToInt32(_view.GVLoaiNganh.GetFocusedRowCellDisplayText("idLoaiNganh"));
+                DialogResult dialogResult = XtraMessageBox.Show("Bạn có chắc chắn muốn xóa?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                if (dialogResult == DialogResult.Yes)
                 {
-                    int id = Convert.ToInt32(_view.GVLoaiNganh.GetFocusedRowCellDisplayText("idLoaiNganh"));
-                    DialogResult dialogResult = XtraMessageBox.Show("Bạn có chắc chắn muốn xóa?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-                    if (dialogResult == DialogResult.Yes)
-                    {
-                        unitOfWorks.LoaiNganhRepository.DeleteById(id);
-                        unitOfWorks.Save();
+                    unitOfWorks.LoaiNganhRepository.DeleteById(id);
+                    if (unitOfWorks.Save())
                         _view.GVLoaiNganh.DeleteRow(row_handle);
-                    }
+                    else
+                        XtraMessageBox.Show("Không thể xóa. Loại Ngành này đang được sử dụng.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
-                else XtraMessageBox.Show("Vui lòng chọn dòng cần xóa.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);                               
             }
-            catch
-            {
-                XtraMessageBox.Show("Không thể xóa. Loại Ngành này đang được sử dụng.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }
+            else
+                XtraMessageBox.Show("Vui lòng chọn dòng cần xóa.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
         }
 
         public void RowIndicator(object sender, RowIndicatorCustomDrawEventArgs e)
